refactor slayer xp drop finder out of slayerplugin file
